+++ src/server/compositor/gl_renderer.h
@@ -37,7 +37,7 @@ public:
     virtual ~GLRenderer() noexcept;
 
     // These are called with a valid GL context:
-    void begin(int rotation) const override;
+    void begin(float rotation) const override;
     void render(CompositingCriteria const& info, graphics::Buffer& buffer) const override;
     void end() const override;
